+++ designer/package.json
@@ -79,7 +79,8 @@
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
     "@types/dagre": "^0.7.44",
-    "@types/hapi": "^18.0.4",
+    "@types/hapi": "^18.0.5",
+    "@types/hapi__vision": "^5.5.2",
     "@types/node": "^14.11.8",
     "@types/react": "^16.9.52",
     "@types/react-dom": "^16.9.8",
