+++ designer/server/plugins/routes/index.ts
@@ -1,0 +1,3 @@
+export * as newConfig from "./newConfig";
+export * as api from "./api";
+export * as configurations from "./configurations";
