+++ package.json
@@ -47,6 +47,7 @@
     "@babel/register": "^7.10.5",
     "@hapi/code": "^8.0.1",
     "@hapi/lab": "^24.0.0",
+    "@types/hapi__vision": "^5.5.2",
     "@typescript-eslint/eslint-plugin": "^4.10.0",
     "@typescript-eslint/parser": "^4.10.0",
     "babel-eslint": "^11.0.0-beta.2",
